test(quick-260803-08o): split module scanner coverage
